refactor code

+++ src/convertToPostgresSQL.js
@@ -15,7 +15,7 @@ export async function convertToPostgresSQL() {
   try {
     let source = buildSourceArray();
     if (source && Array.isArray(source)) {
-      let script = buildScriptInsert(source);
+      let script = buildScriptPostgreSQLScript(source);
       if (script) {
         await logFileWithOuputPath(script, config.outputGenScript);
       }
@@ -29,8 +29,22 @@ export async function convertToPostgresSQL() {
  * build ra script insert dữ liệu
  * @param {array} source input cần build script
  */
-function buildScriptInsert(source) {
+function buildScriptPostgreSQLScript(source) {
   let script = "";
+  if (source && Array.isArray(source)) {
+    let deleteScript = buildDeleteAllScript(source);
+    let arrayScript = [deleteScript];
+    script = arrayScript.join("; \n");
+  }
+  return script;
+}
+
+/**
+ * build ra script delete dữ liệu cũ trước khi insert
+ * @param {array} source input cần build script
+ */
+function buildDeleteAllScript(source) {
+  let deleteScript = "";
   if (
     source &&
     Array.isArray(source) &&
@@ -39,21 +53,21 @@ function buildScriptInsert(source) {
     config.tableName
   ) {
     let allPrimaryValue = source.map((x) => x[config.primaryKeyField]);
-    let tempPrimaryValue = allPrimaryValue[0];
-    let arrayPrimaryDelete = "";
-    if (
-      typeof tempPrimaryValue === "string" ||
-      tempPrimaryValue instanceof String
-    ) {
-      arrayPrimaryDelete = allPrimaryValue.map((x) => `'${x}'`).join(", ");
-    } else {
-      arrayPrimaryDelete = allPrimaryValue.join(", ");
+    if (allPrimaryValue?.length > 0) {
+      let tempPrimaryValue = allPrimaryValue[0];
+      let arrayPrimaryDelete = "";
+      if (
+        typeof tempPrimaryValue === "string" ||
+        tempPrimaryValue instanceof String
+      ) {
+        arrayPrimaryDelete = allPrimaryValue.map((x) => `'${x}'`).join(", ");
+      } else {
+        arrayPrimaryDelete = allPrimaryValue.join(", ");
+      }
+      deleteScript = `delete from ${config.tableName} where ${config.primaryKeyField} in (${arrayPrimaryDelete})`;
     }
-    let deleteScript = `delete from ${config.tableName} where ${config.primaryKeyField} in (${arrayPrimaryDelete})`;
-    let arrayScript = [deleteScript];
-    script = arrayScript.join("; \n");
   }
-  return script;
+  return deleteScript;
 }
 
 /**
